move: OAuth 설정 파일 경로 변경
