+++ src/main/java/com/atwoz/member/application/auth/dto/OAuthTokenResponse.java
@@ -5,13 +5,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
 @NoArgsConstructor
 public class OAuthTokenResponse {
+
     @JsonProperty("access_token")
     private String accessToken;
