+++ src/main/java/com/atwoz/member/config/oauth/OAuthProvider.java
@@ -1,7 +1,7 @@
-package com.atwoz.member.ui.auth.support.oauth;
+package com.atwoz.member.config.oauth;
 
-import com.atwoz.member.ui.auth.support.oauth.OAuthProperties.Provider;
-import com.atwoz.member.ui.auth.support.oauth.OAuthProperties.User;
+import com.atwoz.member.config.oauth.OAuthProperties.Provider;
+import com.atwoz.member.config.oauth.OAuthProperties.User;
 import lombok.Builder;
 import lombok.Getter;
